requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.4
daphne==4.2.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ffmpeg-python==0.2.0
filelock==3.20.3
fsspec==2026.1.0
future==1.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
packaging==26.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
redis==7.1.0
regex==2026.1.15
requests==2.32.5
service-identity==24.2.0
setuptools==80.10.2
sqlparse==0.5.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.2
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
zope.interface==8.2
